--- commons-dao/src/main/java/com/smartsensesolutions/java/commons/filter/sort/SortType.java
@@ -1,33 +1,0 @@
-/*
- * Copyright 2024 smartSense Consulting Solutions Pvt. Ltd
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.smartsensesolutions.java.commons.filter.sort;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/**
- * All supported sort types.
- */
-@AllArgsConstructor
-@Getter
-public enum SortType {
-
-    NONE("none"),
-    ASC("asc"),
-    DESC("desc");
-    private final String value;
-}
