+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/CommonsDaoSampleApplication.java
@@ -1,0 +1,12 @@
+package com.smartsensesolutions.commons.dao.sample;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackages = {"com.smartsensesolutions"})
+public class CommonsDaoSampleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CommonsDaoSampleApplication.class, args);
+    }
+}
