+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/entity/EActivityStatus.java
@@ -1,0 +1,9 @@
+package com.smartsensesolutions.commons.dao.sample.entity;
+
+public enum EActivityStatus {
+    OnLine,
+    NoCall,
+    Away,
+    OffLine,
+    DND
+}
