+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/entity/ELevel.java
@@ -1,0 +1,8 @@
+package com.smartsensesolutions.commons.dao.sample.entity;
+
+public enum ELevel {
+    Noob,
+    Beginner,
+    Intermediate,
+    Expert
+}
