+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/repository/AuthorRepository.java
@@ -1,0 +1,9 @@
+package com.smartsensesolutions.commons.dao.sample.repository;
+
+import com.smartsensesolutions.commons.dao.base.BaseRepository;
+import com.smartsensesolutions.commons.dao.sample.entity.Author;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface AuthorRepository extends BaseRepository<Author, Long> {
+}
