+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/request/AuthorRequest.java
@@ -1,0 +1,6 @@
+package com.smartsensesolutions.commons.dao.sample.request;
+
+import java.util.List;
+
+public record AuthorRequest(String authorName, Long age, List<BookRequest> books) {
+}
