+++ samples/commons-dao-sample/src/main/java/com/smartsensesolutions/commons/dao/sample/request/BookRequest.java
@@ -1,0 +1,4 @@
+package com.smartsensesolutions.commons.dao.sample.request;
+
+public record BookRequest(String bookName, String description) {
+}
