--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/CommonsDaoSampleApplication.java
@@ -1,12 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication(scanBasePackages = {"com.smartsensesolutions", "com.smartsensesolutions.java.commons.dao.sample"})
-public class CommonsDaoSampleApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CommonsDaoSampleApplication.class, args);
-    }
-}
