--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/entity/Address.java
@@ -1,34 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.entity;
-
-import com.smartsensesolutions.java.commons.base.entity.BaseEntity;
-import jakarta.persistence.*;
-import lombok.*;
-import org.hibernate.annotations.CreationTimestamp;
-
-import java.util.Date;
-
-@Entity
-@Table(name = "address")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Address implements BaseEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id", unique = true)
-    private Long id;
-    @Column(name = "city", nullable = false)
-    private String city;
-    @Column(name = "house", nullable = false)
-    private String house;
-    @Column(name = "street", nullable = false)
-    private String street;
-    @Column(name = "active", nullable = false)
-    private boolean active;
-    @Column(name = "created_at", nullable = false)
-    @CreationTimestamp
-    private Date createdAt;
-}
-
