--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/entity/EActivityStatus.java
@@ -1,9 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.entity;
-
-public enum EActivityStatus {
-    OnLine,
-    NoCall,
-    Away,
-    OffLine,
-    DND
-}
