--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/entity/ELevel.java
@@ -1,8 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.entity;
-
-public enum ELevel {
-    Noob,
-    Beginner,
-    Intermediate,
-    Expert
-}
