--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/repository/BooksRepository.java
@@ -1,9 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.repository;
-
-import com.smartsensesolutions.java.commons.base.repository.BaseRepository;
-import com.smartsensesolutions.java.commons.dao.sample.entity.Books;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BooksRepository extends BaseRepository<Books, Long> {
-}
