--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/request/AuthorRequest.java
@@ -1,6 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.request;
-
-import java.util.List;
-
-public record AuthorRequest(String authorName, Long age, List<BookRequest> books) {
-}
