--- samples/commons-dao-sample/src/main/java/com/smartsensesolutions/java/commons/dao/sample/request/BookRequest.java
@@ -1,4 +1,0 @@
-package com.smartsensesolutions.java.commons.dao.sample.request;
-
-public record BookRequest(String bookName, String description) {
-}
